Split block converter tests into multiple test classes
